+++ threedigrid_builder/tests/test_connection_nodes.py
@@ -17,6 +17,7 @@ from threedigrid_builder.grid.connection_nodes import set_bottom_levels
 from threedigrid_builder.grid.connection_nodes import set_calculation_types
 
 import itertools
+import logging
 import numpy as np
 import pygeos
 import pytest
@@ -25,14 +26,14 @@ import pytest
 @pytest.fixture
 def connection_nodes():
     return ConnectionNodes(
-        the_geom=pygeos.points([(0, 0), (10, 0), (10, 20), (10, 30)]),
-        id=np.array([1, 3, 4, 9]),
-        code=np.array(["one", "two", "", ""]),
-        storage_area=np.array([15, 0, np.nan, np.nan]),
-        calculation_type=np.array([1, 2, 5, 2]),
-        bottom_level=np.array([2.1, np.nan, np.nan, np.nan]),
-        manhole_id=[1, -9999, -9999, -9999],
-        drain_level=[1.2, np.nan, np.nan, np.nan],
+        the_geom=pygeos.points([(0, 0), (10, 0), (10, 20), (10, 30), (0, 0)]),
+        id=np.array([1, 3, 4, 9, 10]),
+        code=np.array(["one", "two", "", "", ""]),
+        storage_area=np.array([15, np.nan, np.nan, np.nan, 0]),
+        manhole_id=np.array([42, 11, -9999, -9999, -9999]),
+        calculation_type=np.array([1, 2, 5, 2, 2]),
+        bottom_level=np.array([2.1, np.nan, np.nan, np.nan, 2.1]),
+        drain_level=[1.2, np.nan, np.nan, np.nan, 3.4],
     )
 
 
@@ -42,10 +43,10 @@ def test_get_nodes(connection_nodes):
     nodes = connection_nodes.get_nodes(counter)
     assert isinstance(nodes, Nodes)
 
-    assert_array_equal(nodes.id, [2, 3, 4, 5])
-    assert next(counter) == 6
-    assert_array_equal(nodes.coordinates, [(0, 0), (10, 0), (10, 20), (10, 30)])
-    assert_array_equal(nodes.content_pk, [1, 3, 4, 9])
+    assert_array_equal(nodes.id, [2, 3, 4, 5, 6])
+    assert next(counter) == 7
+    assert_array_equal(nodes.coordinates, [(0, 0), (10, 0), (10, 20), (10, 30), (0, 0)])
+    assert_array_equal(nodes.content_pk, [1, 3, 4, 9, 10])
     assert_array_equal(
         nodes.node_type,
         [
@@ -53,6 +54,7 @@ def test_get_nodes(connection_nodes):
             NodeType.NODE_1D_NO_STORAGE,
             NodeType.NODE_1D_NO_STORAGE,
             NodeType.NODE_1D_NO_STORAGE,
+            NodeType.NODE_1D_STORAGE,
         ],
     )
     assert_array_equal(nodes.calculation_type, connection_nodes.calculation_type)
@@ -207,21 +209,26 @@ def test_bottom_levels_above_invert_level(structure_type, caplog):
     assert_almost_equal(nodes.dmax, [3.0, 20.0])
 
 
-def test_1d2d_properties(connection_nodes):
+def test_1d2d_properties(connection_nodes, caplog):
     """Test setup:
 
-    - node 0: this is a manhole, drain_level is copied
+    - node 0: this has storage_area, drain_level is copied
     - node 1: this is skipped (not in node_idx)
-    - node 2: this is not a manhole, it has 2 channels, lowest bank_level is taken
-    - node 3: this is not a manhole, it has no channels: outcome is NaN
-    - node 4: this is not a manhole, it 1 channel with no bank level: outcome is NaN
+    - node 2: this has no storage_area, it has 2 channels, lowest bank_level is taken
+    - node 3: this has no storage_area, it has no channels: outcome is NaN
+    - node 4: this has no storage_area, it 1 channel with no bank level: outcome is NaN
+    - node 5: this has storage_area == 0, drain_level is copied
     """
-    nodes = Nodes(
-        id=[0, 1, 2, 3, 4],
-        content_pk=[1, 3, 99, 4, 9],
-        dmax=[1.0, 3.0, 2.0, 4.0, 2.3],
+    caplog.set_level(
+        logging.WARNING, logger="threedigrid_builder.grid.connection_nodes"
     )
-    node_idx = np.array([0, 1, 3, 4])
+
+    nodes = Nodes(
+        id=[0, 1, 2, 3, 4, 5],
+        content_pk=[1, 3, 99, 4, 9, 10],
+        dmax=[1.0, 3.0, 2.0, 4.0, 2.3, 5.2],
+    )
+    node_idx = np.array([0, 1, 3, 4, 5])
 
     # channels & cs locations are so that:
     # - channel 32 (CN 1 -> 3, N 0 -> 1), start is 4.0 and end is 0.0
@@ -250,5 +257,10 @@ def test_1d2d_properties(connection_nodes):
         nodes, node_idx, channels, locations
     )
 
-    assert_array_equal(is_closed, [True, False, False, False])
-    assert_array_equal(dpumax, [1.2, 0.0, np.nan, 1.0])
+    assert_array_equal(is_closed, [True, False, False, False, True])
+    assert_array_equal(dpumax, [1.2, 0.0, np.nan, 1.0, 3.4])
+
+    assert (
+        caplog.record_tuples[0][2]
+        == "Manholes [42] have a bottom_level that is above drain_level."
+    )